Move Protocol to type_aliases to avoid circular import
